refactor(diagnostics): complete modular architecture refactoring of diagnostics viewer
This comprehensive refactoring transforms the monolithic diagnostics viewer into a clean,
modular architecture easier to maintain.

## Architectural Transformation
- 1,050 lines → 221 lines in main orchestrator file
- Split into 5 specialized modules with clear separation of concerns
- Thread-safe state management with dependency injection pattern
- Pure rendering logic separated from thread management

Files created:
- `scripts/viewer_components/__init__.py` - Clean module exports
- `scripts/viewer_components/state.py` - Thread-safe state management
- `scripts/viewer_components/renderer.py` - Pure OpenCV rendering logic
- `scripts/viewer_components/threads.py` - Processing and GUI thread coordination
- `scripts/viewer_components/utils.py` - Constants and unified help system

This refactoring establishes a robust foundation for future diagnostics enhancements
while maintaining full backward compatibility with existing functionality.
